+++ ContentPatcher/Framework/Conditions/ConditionType.cs
@@ -137,11 +137,14 @@ namespace ContentPatcher.Framework.Conditions
         /****
         ** String manipulation
         ****/
+        /// <summary>A token which transforms its input text to lowercase.</summary>
+        Lowercase,
+
         /// <summary>A random value selected from the given input.</summary>
         Random,
 
-        /// <summary>A token which transforms its input text to lowercase.</summary>
-        Lowercase,
+        /// <summary>A token which returns a string representation of its input text.</summary>
+        Render,
 
         /// <summary>A token which transforms its input text to uppercase.</summary>
         Uppercase,